fix text wrapping bug in blog posts
removed the CSS rule that applied flexbox to all first paragraphs in blog posts, which was causing weird text wrapping behavior around inline code blocks and links.

the rule was originally added in commit 747386b on nov 17, 2025 to fix badge wrapping on mobile, but it was applying to all first paragraphs even when they didn't contain badges or images.

fixes text wrapping issues:
- after links in 'what is gronka?' post
- before inline code in 'cobalt rate limits' post
- before inline code in 'convert command bug fix' post